pyproject.toml
[tool.poetry]
name = "fastapi-project"
version = "0.1.0"
description = ""
authors = ["Aldirane <aldirane@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.13"
fastapi = "^0.103.2"
alembic = "^1.12.0"
uvicorn = "^0.23.2"
pytest-asyncio = "^0.21.1"
asyncpg = "^0.28.0"
dataclass-factory = "^2.16"
httpx = "^0.25.0"
requests = "^2.31.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {version = "~2.1", extras = ["email"]}
python-dateutil = "^2.8.2"
pandas = "^2.1.1"
pyyaml = "^6.0.1"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
